refactor: . move Occurrence to a separate file
