+++ src/Suggestor/Suggestor.ts
@@ -2,7 +2,8 @@ import type { Editor, EditorPosition } from 'obsidian';
 import type { Settings } from '../Config/Settings';
 import { DateParser } from '../Query/DateParser';
 import { doAutocomplete } from '../lib/DateAbbreviations';
-import { Occurrence, Recurrence } from '../Task/Recurrence';
+import { Occurrence } from '../Task/Occurrence';
+import { Recurrence } from '../Task/Recurrence';
 import {
     type DefaultTaskSerializerSymbols,
     allTaskPluginEmojis,
